сделал баг фикс + попытался сделать автозап. слаг
