+++ L23/DjangoProject/HW23/apps.py
@@ -4,3 +4,6 @@ from django.apps import AppConfig
 class Hw23Config(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'HW23'
+
+    def ready(self):
+        from . import signals
